Revamp dead man switch model
(DMS = Dead Man Switch)

- Start a DMS session which is disposable (and has a CancellationTokenSource)
- Separate the notifications of the DMS from the DMS watcher
- Separate the triggering of the DMS from the DMS watcher
- Rename Task to Worker and let the worker return TResult

The intention is to optimize and simplify the DMS tasks, and to let workers return results.
I also wish to change the 'Delay' property on a worker, the worker should return the 'NextDelay' instead of changing a property value. This makes it a lot more intuitive.
Lots of work left to do though.